Refactored video call screen and updated components export
