+++ template/fpe-api/components.ts
@@ -1,13 +1,22 @@
-//Precall components
-export * from '../src/components/precall/index';
-//video call components
-export * from '../src/pages/video-call/index';
 //Icons
 export {default as icons} from '../src/assets/icons';
-//Settings
+
+//create screen
+export {default as Create} from '../src/pages/Create';
+
+//share screen
+export {default as Share} from '../src/components/Share';
+
+//Precall components
+export * from '../src/components/precall/index';
+
+//video call components
+export * from '../src/pages/video-call/index';
+export {default as GridLayout} from '../src/components/GridVideo';
+export {default as PinnedLayout} from '../src/components/PinnedVideo';
+export {default as VideoComponentWithDefaultLayout} from '../src/pages/video-call/VideoComponent';
+export {MaxVideoRenderer as VideoContentComponent} from '../src/pages/video-call/VideoRenderer';
+
+//Settings screen
 export {default as HostControlView} from '../src/components/HostControlView';
 export {default as SelectDevice} from '../src/subComponents/SelectDevice';
-//Video Rendering Component
-export {default as MaxVideoView} from '../agora-rn-uikit/src/Views/MaxVideoView';
-export {default as Create} from '../src/pages/Create';
-export {default as Share} from '../src/components/Share';
